Use all of ActiveRecord::AttributeMethods
