+++ lib/superstore/attribute_methods/primary_key.rb
@@ -4,7 +4,7 @@ module Superstore
       extend ActiveSupport::Concern
 
       included do
-        include ActiveRecord::AttributeMethods::PrimaryKey
+        # include ActiveRecord::AttributeMethods::PrimaryKey
         attribute :id, type: :string
 
         extend ClassOverrides
